Exposing the VariableSubstitutionSqlParser for extensibility
